+++ example/app/web/flutter_service_worker.js
@@ -3,11 +3,11 @@ const MANIFEST = 'flutter-app-manifest';
 const TEMP = 'flutter-temp-cache';
 const CACHE_NAME = 'flutter-app-cache';
 
-const RESOURCES = {"flutter_bootstrap.js": "ff59af68a1f02c99eddecfd3427552d4",
+const RESOURCES = {"flutter_bootstrap.js": "7fd0fe9c13091bcbae14373df0ca0972",
 "version.json": "11e1a177d57c5da2bc6e3a6acc3010f4",
 "index.html": "586ac9a167f4eb9615dadda7d1131ef7",
 "/": "586ac9a167f4eb9615dadda7d1131ef7",
-"main.dart.js": "184d6b3b069e8db0dd3b3483ae22a5da",
+"main.dart.js": "2967b13b9a21053e0f19f5047c039519",
 "flutter.js": "f393d3c16b631f36852323de8e583132",
 "favicon.png": "5dcef449791fa27946b3d35ad8803796",
 "icons/Icon-192.png": "ac9a721a12bbc803b44f645561ecb1e1",